requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-spectacular>=0.24.0,<0.25
flake8>=4.0.1,<4.1
psycopg2>=2.9.3,<2.10
sqlparse==0.4.3
tzdata==2023.3
uwsgi>=2.0.20,<2.1
